requirements.txt
google-api-python-client==2.148.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
rich==13.9.2
